Tests/LibWeb/Text/expected/DOM/getElementsByClassName-multiple.txt

document.getElementsByClassName("te st").length: 3
<DIV id="1" >
<DIV id="2" >
<DIV id="3" >
